+++ src/entities/graph/Vertex.java
@@ -12,7 +12,7 @@ public class Vertex {
 	
 	private int m_Id;
 	
-	private float m_Distance = Float.MAX_VALUE;
+	private int m_Distance = Integer.MAX_VALUE;
 	
 	private List<Edge> m_Edges;
 	
@@ -24,9 +24,9 @@ public class Vertex {
 	
 	/**
 	 * Get distance to a specific vertex depending on the used algorithm.
-	 * @return Distance as float value, initialized with Float.MAX_VALUE
+	 * @return Distance as int value, initialized with Float.MAX_VALUE
 	 */
-	public float GetDistance() {
+	public int GetDistance() {
 		return m_Distance;
 	}
 	
@@ -44,7 +44,7 @@ public class Vertex {
 	 * Set distance to a specific vertex depending on the used algorithm.
 	 * @param _distance The new distance
 	 */
-	public void SetDistance(float _distance) {
+	public void SetDistance(int _distance) {
 		m_Distance = _distance;
 	}
 	
@@ -108,12 +108,12 @@ public class Vertex {
 	 * @param _vertex Vertex to get distance to
 	 * @return Distance to vertex, 0 if vertex is the same, Float.MAX_VALUE if vertices aren't neighbors
 	 */
-	public float GetDistanceToVertex(Vertex _vertex) {
+	public int GetDistanceToVertex(Vertex _vertex) {
 		
 		if(this.equals(_vertex))
-			return 0f;
+			return 0;
 		
-		float distance = Float.MAX_VALUE;
+		int distance = Integer.MAX_VALUE;
 		
 		for(Edge edge : this.GetEdges()) {
 			if(edge.GetNeighbour(this).equals(_vertex)) {
